new/requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning
tensorflow>=2.8.0
scikit-learn>=1.0.0

# Data handling
h5py>=3.6.0
tables>=3.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: GPU acceleration
# tensorflow-gpu>=2.8.0  # Uncomment if using GPU